Move data saver to the expected file and include some log traces
